backEnd/requirements.txt
fastapi[all]
faker
decouple-python
sentence-transformers
pydantic
mysql-connector-python
openai